requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-redis==6.0.0
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
numpy==2.3.2
portalocker==3.2.0
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
python-decouple==3.8
pywin32==311
qdrant-client==1.15.0
redis==6.2.0
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
